refactored handlers into handle package
