+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/UserService.java
@@ -1,10 +1,8 @@
 package edu.byu.cs.tweeter.client.model.service;
 
-import android.os.Handler;
-import android.os.Message;
-import androidx.annotation.NonNull;
 import edu.byu.cs.tweeter.client.cache.Cache;
 import edu.byu.cs.tweeter.client.model.service.backgroundTask.*;
+import edu.byu.cs.tweeter.client.model.service.handler.*;
 import edu.byu.cs.tweeter.client.presenter.MainActivityPresenter;
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
@@ -28,32 +26,6 @@ public class UserService {
         executor.execute(getUserTask);
     }
 
-    /**
-     * Message handler (i.e., observer) for GetUserTask.
-     */
-    private class GetUserHandler extends Handler {
-        private GetUserObserver observer;
-
-        public GetUserHandler(GetUserObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(GetUserTask.SUCCESS_KEY);
-            if (success) {
-                User user = (User) msg.getData().getSerializable(GetUserTask.USER_KEY);
-                observer.handleSuccess(user);
-            } else if (msg.getData().containsKey(GetUserTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(GetUserTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(GetUserTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(GetUserTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
-    }
-
     public interface LoginObserver {
         void handleSuccess(User loggedInUser, AuthToken authToken);
 
@@ -69,35 +41,6 @@ public class UserService {
                 new LoginHandler(loginObserver));
         ExecutorService executor = Executors.newSingleThreadExecutor();
         executor.execute(loginTask);
-    }
-
-    /**
-     * Message handler (i.e., observer) for LoginTask
-     */
-    private class LoginHandler extends Handler {
-        LoginObserver observer;
-
-        public LoginHandler(LoginObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(LoginTask.SUCCESS_KEY);
-            if (success) {
-                User loggedInUser = (User) msg.getData().getSerializable(LoginTask.USER_KEY);
-                AuthToken authToken = (AuthToken) msg.getData().getSerializable(LoginTask.AUTH_TOKEN_KEY);
-
-                observer.handleSuccess(loggedInUser, authToken);
-
-            } else if (msg.getData().containsKey(LoginTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(LoginTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(LoginTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(LoginTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
     }
 
     public interface RegisterObserver {
@@ -116,30 +59,6 @@ public class UserService {
         executor.execute(registerTask);
     }
 
-    private class RegisterHandler extends Handler {
-        RegisterObserver observer;
-
-        public RegisterHandler(RegisterObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(RegisterTask.SUCCESS_KEY);
-            if (success) {
-                User registeredUser = (User) msg.getData().getSerializable(RegisterTask.USER_KEY);
-                AuthToken authToken = (AuthToken) msg.getData().getSerializable(RegisterTask.AUTH_TOKEN_KEY);
-                observer.handleSuccess(registeredUser, authToken);
-            } else if (msg.getData().containsKey(RegisterTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(RegisterTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(RegisterTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(RegisterTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
-    }
-
     public interface LogoutObserver {
         void handleSuccess();
 
@@ -154,29 +73,6 @@ public class UserService {
         executor.execute(logoutTask);
     }
 
-    // LogoutHandler
-    private class LogoutHandler extends Handler {
-        private LogoutObserver observer;
-
-        public LogoutHandler(LogoutObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(LogoutTask.SUCCESS_KEY);
-            if (success) {
-                observer.handleSuccess();
-            } else if (msg.getData().containsKey(LogoutTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(LogoutTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(LogoutTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(LogoutTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
-    }
-
     public interface GetFollowersCountObserver {
         void handleSuccess(int count);
 
@@ -189,32 +85,8 @@ public class UserService {
         ExecutorService executor = Executors.newSingleThreadExecutor();
         // Get count of most recently selected user's followers.
         GetFollowersCountTask followersCountTask = new GetFollowersCountTask(Cache.getInstance().getCurrUserAuthToken(),
-                selectedUser, new UserService.GetFollowersCountHandler(getFollowersCountObserver));
+                selectedUser, new GetFollowersCountHandler(getFollowersCountObserver));
         executor.execute(followersCountTask);
-    }
-
-    // GetFollowersCountHandler
-    private class GetFollowersCountHandler extends Handler {
-        UserService.GetFollowersCountObserver observer;
-
-        public GetFollowersCountHandler(UserService.GetFollowersCountObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(GetFollowersCountTask.SUCCESS_KEY);
-            if (success) {
-                int count = msg.getData().getInt(GetFollowersCountTask.COUNT_KEY);
-                observer.handleSuccess(count);
-            } else if (msg.getData().containsKey(GetFollowersCountTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(GetFollowersCountTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(GetFollowersCountTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(GetFollowersCountTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
     }
 
     public interface GetFollowingCountObserver {
@@ -229,32 +101,8 @@ public class UserService {
         ExecutorService executor = Executors.newSingleThreadExecutor();
         // Get count of most recently selected user's followees (who they are following)
         GetFollowingCountTask followingCountTask = new GetFollowingCountTask(Cache.getInstance().getCurrUserAuthToken(),
-                selectedUser, new UserService.GetFollowingCountHandler(getFollowingCountObserver));
+                selectedUser, new GetFollowingCountHandler(getFollowingCountObserver));
         executor.execute(followingCountTask);
-    }
-
-    // GetFollowingCountHandler
-    private class GetFollowingCountHandler extends Handler {
-        UserService.GetFollowingCountObserver observer;
-
-        GetFollowingCountHandler(UserService.GetFollowingCountObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(GetFollowingCountTask.SUCCESS_KEY);
-            if (success) {
-                int count = msg.getData().getInt(GetFollowingCountTask.COUNT_KEY);
-                observer.handleSuccess(count);
-            } else if (msg.getData().containsKey(GetFollowingCountTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(GetFollowingCountTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(GetFollowingCountTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(GetFollowingCountTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
     }
 
     public interface IsFollowerObserver {
@@ -268,32 +116,9 @@ public class UserService {
 
     public void isFollower(User selectedUser, MainActivityPresenter.IsFollowerObserver isFollowerObserver) {
         IsFollowerTask isFollowerTask = new IsFollowerTask(Cache.getInstance().getCurrUserAuthToken(),
-                Cache.getInstance().getCurrUser(), selectedUser, new UserService.IsFollowerHandler(isFollowerObserver));
+                Cache.getInstance().getCurrUser(), selectedUser, new IsFollowerHandler(isFollowerObserver));
         ExecutorService executor = Executors.newSingleThreadExecutor();
         executor.execute(isFollowerTask);
     }
 
-    // IsFollowerHandler
-    private class IsFollowerHandler extends Handler {
-        UserService.IsFollowerObserver observer;
-
-        IsFollowerHandler(UserService.IsFollowerObserver observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void handleMessage(@NonNull Message msg) {
-            boolean success = msg.getData().getBoolean(IsFollowerTask.SUCCESS_KEY);
-            if (success) {
-                boolean isFollower = msg.getData().getBoolean(IsFollowerTask.IS_FOLLOWER_KEY);
-                observer.handleSuccess(isFollower);
-            } else if (msg.getData().containsKey(IsFollowerTask.MESSAGE_KEY)) {
-                String message = msg.getData().getString(IsFollowerTask.MESSAGE_KEY);
-                observer.handleMessage(message);
-            } else if (msg.getData().containsKey(IsFollowerTask.EXCEPTION_KEY)) {
-                Exception ex = (Exception) msg.getData().getSerializable(IsFollowerTask.EXCEPTION_KEY);
-                observer.handleException(ex);
-            }
-        }
-    }
 }
