+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/GetFeedHandler.java
@@ -1,0 +1,38 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+import androidx.annotation.NonNull;
+import edu.byu.cs.tweeter.client.model.service.StatusService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.GetFeedTask;
+import edu.byu.cs.tweeter.model.domain.Status;
+
+import java.util.List;
+
+/**
+ * Message handler (i.e., observer) for GetFeedTask.
+ */
+public class GetFeedHandler extends Handler {
+    private StatusService.GetFeedObserver observer;
+
+    public GetFeedHandler(StatusService.GetFeedObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+
+        boolean success = msg.getData().getBoolean(GetFeedTask.SUCCESS_KEY);
+        if (success) {
+            List<Status> statuses = (List<Status>) msg.getData().getSerializable(GetFeedTask.STATUSES_KEY);
+            boolean hasMorePages = msg.getData().getBoolean(GetFeedTask.MORE_PAGES_KEY);
+            observer.handleSuccess(statuses, hasMorePages);
+        } else if (msg.getData().containsKey(GetFeedTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(GetFeedTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(GetFeedTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(GetFeedTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
