+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/GetFollowersCountHandler.java
@@ -1,0 +1,31 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+import androidx.annotation.NonNull;
+import edu.byu.cs.tweeter.client.model.service.UserService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.GetFollowersCountTask;
+
+// GetFollowersCountHandler
+public class GetFollowersCountHandler extends Handler {
+    UserService.GetFollowersCountObserver observer;
+
+    public GetFollowersCountHandler(UserService.GetFollowersCountObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(GetFollowersCountTask.SUCCESS_KEY);
+        if (success) {
+            int count = msg.getData().getInt(GetFollowersCountTask.COUNT_KEY);
+            observer.handleSuccess(count);
+        } else if (msg.getData().containsKey(GetFollowersCountTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(GetFollowersCountTask.MESSAGE_KEY);
+            observer.handleMessage(message);
+        } else if (msg.getData().containsKey(GetFollowersCountTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(GetFollowersCountTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
