+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/GetFollowersHandler.java
@@ -1,0 +1,39 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+import androidx.annotation.NonNull;
+import edu.byu.cs.tweeter.client.model.service.FollowService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.GetFollowersTask;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.PagedTask;
+import edu.byu.cs.tweeter.model.domain.User;
+
+import java.util.List;
+
+/**
+ * Message handler (i.e., observer) for GetFollowersTask.
+ */
+public class GetFollowersHandler extends Handler {
+    private FollowService.GetFollowersObserver observer;
+
+    public GetFollowersHandler(FollowService.GetFollowersObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+
+        boolean success = msg.getData().getBoolean(GetFollowersTask.SUCCESS_KEY);
+        if (success) {
+            List<User> followers = (List<User>) msg.getData().getSerializable(PagedTask.ITEMS_KEY);
+            boolean hasMorePages = msg.getData().getBoolean(GetFollowersTask.MORE_PAGES_KEY);
+            observer.handleSuccess(followers, hasMorePages);
+        } else if (msg.getData().containsKey(GetFollowersTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(GetFollowersTask.MESSAGE_KEY);
+            observer.handleFailure(message);
+        } else if (msg.getData().containsKey(GetFollowersTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(GetFollowersTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
