+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/handler/IsFollowerHandler.java
@@ -1,0 +1,31 @@
+package edu.byu.cs.tweeter.client.model.service.handler;
+
+import android.os.Handler;
+import android.os.Message;
+import androidx.annotation.NonNull;
+import edu.byu.cs.tweeter.client.model.service.UserService;
+import edu.byu.cs.tweeter.client.model.service.backgroundTask.IsFollowerTask;
+
+// IsFollowerHandler
+public class IsFollowerHandler extends Handler {
+    UserService.IsFollowerObserver observer;
+
+    public IsFollowerHandler(UserService.IsFollowerObserver observer) {
+        this.observer = observer;
+    }
+
+    @Override
+    public void handleMessage(@NonNull Message msg) {
+        boolean success = msg.getData().getBoolean(IsFollowerTask.SUCCESS_KEY);
+        if (success) {
+            boolean isFollower = msg.getData().getBoolean(IsFollowerTask.IS_FOLLOWER_KEY);
+            observer.handleSuccess(isFollower);
+        } else if (msg.getData().containsKey(IsFollowerTask.MESSAGE_KEY)) {
+            String message = msg.getData().getString(IsFollowerTask.MESSAGE_KEY);
+            observer.handleMessage(message);
+        } else if (msg.getData().containsKey(IsFollowerTask.EXCEPTION_KEY)) {
+            Exception ex = (Exception) msg.getData().getSerializable(IsFollowerTask.EXCEPTION_KEY);
+            observer.handleException(ex);
+        }
+    }
+}
